#37 rename platform io environment, split at serial command into seperate header files
