+++ platformio.ini
@@ -8,7 +8,7 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
-[env:esp32cam]
+[env:esp32CarWifiWithCamera]
 platform = espressif32
 board = esp32cam
 board_build.partitions = partitions/4mb_spiffs.csv
@@ -19,7 +19,7 @@ board_build.mcu = esp32
 framework = arduino
 monitor_speed = 115200
 
-build_src_filter = +<*.h> +<carWithCamera/*.h> +<carWithCamera/main.cpp>
+build_src_filter = +<*.h> +<esp32CarWifiWithCamera/*.h> +<esp32CarWifiWithCamera/main.cpp>
 lib_deps = 
 	ESP32Async/AsyncTCP
 	ESP32Async/ESPAsyncWebServer
@@ -29,12 +29,26 @@ lib_deps =
 ;; when pushing firmware, short GPIO0 and GND pin
 ;; if firmware upload fails, then push the boot button once
 
-[env:nodemcu-32s-carBLE]
+[env:esp32CarBleWithSteering]
 platform = espressif32
 board = upesy_wroom
 framework = arduino
 monitor_speed = 115200
-build_src_filter = +<*.h> +<bleCar/*.h> +<bleCar/main.cpp>
+build_src_filter = +<*.h> +<esp32CarBleWithSteering/*.h> +<esp32CarBleWithSteering/main.cpp>
+lib_deps = 
+	cytrontechnologies/Cytron Motor Drivers Library@^1.0.1
+	madhephaestus/ESP32Servo@^3.0.9
+	waspinator/AccelStepper@^1.64
+
+[env:basicBleWifi]
+;; To check basic functionality of BLE, Wifi Async Server
+;; debug mode
+;; Check save preference on reload
+platform = espressif32
+board = upesy_wroom
+framework = arduino
+monitor_speed = 115200
+build_src_filter = +<*.h> +<basicBleWifi/*.h> +<basicBleWifi/main.cpp>
 lib_deps = 
 	cytrontechnologies/Cytron Motor Drivers Library@^1.0.1
 	madhephaestus/ESP32Servo@^3.0.9
