+++ src/basicBleWifi/README.md
@@ -1,0 +1,4 @@
+## Basic Ble Wifi
+
+* To Check basic functioanlity of BLE and Wifi Async library
+* To check AT commands, save feature
